bots/movie-game-bot/requirements.txt
bubbletea-chat
openai
python-dotenv
